[ADD] CompanyPagerViewController to contains related vcs
