+++ Mearcats/Storyboard/Company.storyboard
@@ -16,7 +16,7 @@
                     <view key="view" contentMode="scaleToFill" id="CYN-85-xof">
                         <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
+                        <color key="backgroundColor" white="0.66666666666666663" alpha="1" colorSpace="calibratedWhite"/>
                         <viewLayoutGuide key="safeArea" id="Pie-Jh-6hm"/>
                     </view>
                 </viewController>
